rename datasources module to datasource
