+++ hvpy/__init__.py
@@ -1,5 +1,5 @@
 from .facade import *
 from .config import set_api_url
 from .version import __version__
-from .datasources import *
+from .datasource import *
 from .utils import create_layers
